fix(io,client,config,graphics,world): Packet.G1B returns int8 to match Java byte signedness
The Java g1b() returns a signed byte (-128..127), promoted to int via sign
extension. The Go port returned Go byte (uint8 0..255), so int(p.G1B())
zero-extended — silently corrupting any negative wire value (0x80..0xFF
became 128..255 instead of -128..-1).

The CRITICAL site was client.go opcode 150 (varp set), where the varp
value is signed in Java and the prior Go widening lost the sign.
ReadZonePacket opcodes 69/23/151 had the same flaw for negative tile
deltas. Other call sites (font masks, palette indices, tile overlay IDs)
store the raw bit pattern without semantic widening; those keep their
byte-typed storage via an explicit bit-preserving cast at the assignment.

Field-type follow-ups (LocType.Ambient/Contrast, NpcType.Size,
ObjType.ManWearOffsetY/WomanWearOffsetY) still hold the byte type;
those values participate in arithmetic later and should be int8 for
full Java fidelity, but the immediate CRITICAL is the wire-protocol
sign-extension and that is what this commit lands.

+++ pkg/jagex2/io/packet.go
@@ -201,8 +201,12 @@ func (p *Packet) G1() int {
 	return n
 }
 
-func (p *Packet) G1B() byte {
-	n := p.Data[p.Pos]
+// G1B reads a signed byte (Java byte; -128..127). Use int(p.G1B()) when
+// widening — Go's int(int8) sign-extends, matching Java's byte->int
+// promotion. For unsigned storage of the raw byte pattern, wrap as
+// byte(p.G1B()) (this is a bit-preserving reinterpretation).
+func (p *Packet) G1B() int8 {
+	n := int8(p.Data[p.Pos])
 	p.Pos++
 	return n
 }
